fix: 콕찌르기 API 경로 수정
