+++ src/core/api/functions/notificationAPI.ts
@@ -5,6 +5,11 @@ const notificationAPI = {
     return await baseInstance.post(`/notifications`, null, {
       params
     });
+  },
+  getMemberPoke: async (roomId: string, memberId: string) => {
+    return await baseInstance.get(
+      `/notifications/rooms/${roomId}/members/${memberId}`
+    );
   }
 };
